+++ spec/gruf/newrelic/client_interceptor_spec.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Copyright (c) 2018-present, BigCommerce Pty. Ltd. All rights reserved
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
@@ -14,37 +12,34 @@
 # WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-require 'new_relic/agent'
-require 'new_relic/agent/distributed_tracing'
+#
+require 'spec_helper'
 
-module Gruf
-  module Newrelic
-    ##
-    # New Relic transaction tracing for Gruf endpoints
-    #
-    class ServerInterceptor < ::Gruf::Interceptors::ServerInterceptor
-      include ::NewRelic::Agent::Instrumentation::ControllerInstrumentation
+describe Gruf::Newrelic::ClientInterceptor do
+  let(:interceptor) { described_class.new }
+  let(:nr_tracer) { ::NewRelic::Agent::DistributedTracing }
+  let(:nr_header_data) { 'newrelic-header-data' }
+  let(:nr_trace_instance) { instance_double(::NewRelic::Agent::DistributedTracePayload, http_safe: nr_header_data) }
+  let(:request_context) { double(:request_context, metadata: {}) }
+  subject { interceptor.call(request_context: request_context){ true } }
 
-      def call
-        opts = {
-          category: Gruf::Newrelic.server_category,
-          class_name: request.service,
-          name: request.method_key
-        }
+  describe '.call' do
+    it 'yields' do
+      expect(interceptor).to receive(:call).once.and_yield
+      subject
+    end
 
-        # Yield to the given block with NewRelic tracing.
-        # http://www.rubydoc.info/github/newrelic/rpm/NewRelic%2FAgent%2FInstrumentation%2FControllerInstrumentation:perform_action_with_newrelic_trace
-        perform_action_with_newrelic_trace(opts) do
-          accept_distributed_tracing
-          yield
-        end
-      end
+    it 'adds nr tracing header' do
+      expect(nr_tracer).to receive(:create_distributed_trace_payload).once.and_return(nr_trace_instance)
+      subject
+      expect(request_context.metadata[Gruf::Newrelic::NEWRELIC_TRACE_HEADER]).to eq(nr_header_data)
+    end
 
-      private
-
-      def accept_distributed_tracing
-        payload = request.active_call.metadata[NEWRELIC_TRACE_HEADER]
-        ::NewRelic::Agent::DistributedTracing.accept_distributed_trace_payload(payload)
+    context "no header" do
+      it 'skips nr tracing header' do
+        expect(nr_tracer).to receive(:create_distributed_trace_payload).once.and_return(nil)
+        subject
+        expect(request_context.metadata[Gruf::Newrelic::NEWRELIC_TRACE_HEADER]).to eq(nil)
       end
     end
   end